separate editor script to fix build
