+++ ProjectSettings/ProjectSettings.asset
@@ -144,6 +144,8 @@ PlayerSettings:
   - {fileID: 11400000, guid: cbb1003a409a8463b828e0a7228c0d27, type: 2}
   - {fileID: 0}
   - {fileID: 0}
+  - {fileID: -4510615847915551955, guid: f66d99971f7d64b64af04027f9be353c, type: 2}
+  - {fileID: -8816637224271811695, guid: 2039691138b3a467c80f76925e6fa8e4, type: 2}
   metroInputSource: 0
   wsaTransparentSwapchain: 0
   m_HolographicPauseOnTrackingLoss: 1
@@ -559,7 +561,7 @@ PlayerSettings:
   enableCrashReportAPI: 0
   cameraUsageDescription: 
   locationUsageDescription: 
-  microphoneUsageDescription: 
+  microphoneUsageDescription: N/A
   bluetoothUsageDescription: 
   macOSTargetOSVersion: 10.13.0
   switchNMETAOverride:
